requirements.txt
Flask>=2.2,<3.0
Flask-SocketIO>=5.3.0
paho-mqtt>=1.6
python-socketio>=5.9.0
